backend/requirements.txt
# Add your dependencies here
python-dotenv==1.1.1
google-genai==1.30.0
uvicorn==0.35.0
fastapi==0.116.1
